18.08 2026 changes

+++ services/ILeaveCalculationService.cs
@@ -4,8 +4,5 @@ namespace ErpPersonelLeaveSystem.Services;
 
 public interface ILeaveCalculationService
 {
-    PayrollCalculationResult CalculatePayroll(decimal monthlySalary, LeaveType leaveType, int leaveDays);
+    PayrollCalculationResult CalculatePayroll(decimal monthlySalary, LeaveType leaveType, decimal leaveDays);
 }
-
-//dışarıdan monthlysalary, leavetype, leavedays parametreleri alınır, calculatepayroll fonksiyonu hesaplamayı yapıp geriye net maaşı ve kesinti miktarını ( payrollcalculationresult ) geri döndürür.
-// burada CalculatePayroll fonksiyonunu Interface içerisinden import ettik
